+++ internal/ui/tree.go
@@ -170,23 +170,23 @@ func (t *Tree) buildHeader() *gtk.Box {
 	t.countLabel.SetXAlign(0)
 	top.Append(t.countLabel)
 
-	sortBtn := gtk.NewButtonFromIconName("view-sort-descending-symbolic")
+	sortBtn := gtk.NewButtonFromIconName("atlas-sort-symbolic")
 	sortBtn.AddCSSClass("flat")
 	sortBtn.SetTooltipText("Sort by name")
 	sortBtn.ConnectClicked(func() {
 		t.sortRecent = !t.sortRecent
 		if t.sortRecent {
 			sortBtn.SetTooltipText("Sort by last edited")
-			sortBtn.SetIconName("document-open-recent-symbolic")
+			sortBtn.SetIconName("atlas-recent-symbolic")
 		} else {
 			sortBtn.SetTooltipText("Sort by name")
-			sortBtn.SetIconName("view-sort-descending-symbolic")
+			sortBtn.SetIconName("atlas-sort-symbolic")
 		}
 		t.Refresh()
 	})
 	top.Append(sortBtn)
 
-	newBtn := gtk.NewButtonFromIconName("list-add-symbolic")
+	newBtn := gtk.NewButtonFromIconName("atlas-add-symbolic")
 	newBtn.AddCSSClass("flat")
 	newBtn.SetTooltipText("New note here (Ctrl+N)")
 	newBtn.ConnectClicked(func() { t.promptNewNote(t.SelectedFolder()) })
@@ -213,7 +213,7 @@ func (t *Tree) buildEmptyState() *gtk.Box {
 	box.SetVExpand(true)
 	box.SetVisible(false)
 
-	icon := gtk.NewImageFromIconName("folder-symbolic")
+	icon := gtk.NewImageFromIconName("atlas-folder-symbolic")
 	icon.SetPixelSize(28)
 	icon.AddCSSClass("dim-label")
 	box.Append(icon)